10/17 afternoon

+++ routes/web.php
@@ -8,39 +8,35 @@ use App\Models\Product;
 use Illuminate\Support\Facades\Route;
 use Illuminate\Database\Eloquent\toSql;
 
-Route::get('/login', function() {
+Route::get('/login', function () {
     return view('dashboard.login');
 });
 
 Route::post('/admin/login', [DashboardController::class, 'login']);
-
-// Route::get('/admin/registration', function() {
-//     return view('dashboard.registration');
-// });
-
 Route::get('/admin/registration', [RegisterController::class, 'registration']);
 Route::post('/admin/store', [RegisterController::class, 'store']);
+Route::post('/admin/review', [DashboardController::class, 'adminReview']);
 
+Route::get('/admin', [DashboardController::class, 'admin'])->middleware('admin')->middleware('admin');
+Route::get('/adminSessionDestroy', [DashboardController::class, 'adminSessionDestroy'])->middleware('admin');
+Route::get('/admin/orders', [DashboardController::class, 'orders'])->middleware('admin');
+Route::get('/admin/orderDetail/{id}', [DashboardController::class, 'orderDetail'])->middleware('admin');
+Route::patch('/admin/orderUpdate/{id}', [DashboardController::class, 'orderUpdate'])->middleware('admin');
+Route::get('/admin/logout', [DashboardController::class, 'logout'])->middleware('admin');
 
-Route::get('/admin', [DashboardController::class, 'admin'])->middleware('admin');
-Route::get('/adminSessionDestroy', [DashboardController::class, 'adminSessionDestroy']);
-Route::get('/admin/orders', [DashboardController::class, 'orders']);
-Route::get('/admin/orderDetail/{id}', [DashboardController::class, 'orderDetail']);
-Route::patch('/admin/orderUpdate/{id}', [DashboardController::class, 'orderUpdate']);
-Route::get('/admin/logout', [DashboardController::class, 'logout']);
-
-Route::post('/admin/forgotPassword', [DashboardController::class, 'forgotPassword']);
-Route::get('/admin/forgotPassword', function() {
+Route::post('/admin/forgotPassword', [DashboardController::class, 'forgotPassword'])->middleware('admin');
+Route::get('/admin/forgotPassword', function () {
     return view('dashboard.forgotPassword');
-});
+})->middleware('admin');
 
-Route::get('/admin/products', [DashboardController::class, 'products']);
-Route::get('/admin/productsEdit/{id}', [DashboardController::class, 'productsEdit']);
-Route::post('/admin/productsUpdate/{id}', [DashboardController::class, 'productsUpdate']);
-Route::get('/admin/addProduct', [DashboardController::class, 'addProduct']);
-Route::post('/admin/addnewProduct', [DashboardController::class, 'addnewProduct']);
+Route::get('/admin/products', [DashboardController::class, 'products'])->middleware('admin');
+Route::get('/admin/productsEdit/{id}', [DashboardController::class, 'productsEdit'])->middleware('admin');
+Route::post('/admin/productsUpdate/{id}', [DashboardController::class, 'productsUpdate'])->middleware('admin');
+Route::get('/admin/addProduct', [DashboardController::class, 'addProduct'])->middleware('admin');
+Route::post('/admin/addnewProduct', [DashboardController::class, 'addnewProduct'])->middleware('admin');
 
-Route::delete('admin/productDestroy/{post}', [DashboardController::class, 'productDestroy']);
+
+Route::delete('admin/productDestroy/{post}', [DashboardController::class, 'productDestroy'])->middleware('admin');
 
 // products
 Route::get('/', [ProductController::class, 'index']);
@@ -55,8 +51,7 @@ Route::get('contactus', [ProductController::class, 'contactus']);
 Route::post('contactus', [ProductController::class, 'contactuspost']);
 
 Route::get('/session', function () {
-    if(session()->has('cart'))
-    {
+    if (session()->has('cart')) {
         $count = count(session('cart'));
         echo $count . "<br>";
         print_r(session()->all());
@@ -66,4 +61,3 @@ Route::get('/session', function () {
 });
 
 Route::get('/sessionDestroy', [ProductController::class, 'sessionDestroy']);
-
